Revert to 2a765ae

+++ src/utils/analysis_dustywind.F90
@@ -2,7 +2,7 @@
 ! The Phantom Smoothed Particle Hydrodynamics code, by Daniel Price et al. !
 ! Copyright (c) 2007-2023 The Authors (see AUTHORS)                        !
 ! See LICENCE file for usage and distribution conditions                   !
-! http://phantomsph.github.io/                                             !
+! http://phantomsph.bitbucket.io/                                          !
 !--------------------------------------------------------------------------!
 module analysis
 !
@@ -24,8 +24,8 @@ module analysis
 
  private
  integer, parameter :: N = 1024 !32
- real, parameter :: theta = 0., phi = 0.
- real, parameter :: u(3) = (/ sin(theta)*cos(phi), sin(theta)*sin(phi), cos(theta) /)
+ double precision, parameter :: theta = 0., phi = 0.
+ double precision, parameter :: u(3) = (/ sin(theta)*cos(phi), sin(theta)*sin(phi), cos(theta) /)
 
 contains
